requirements.txt
PyJWT==2.6.0
bcrypt==4.0.1
httpx==0.23.3
passlib==1.7.4
sqlmodel==0.0.8
fastapi==0.93.0
uvicorn==0.20.0
pydantic==1.10.5
starlette==0.25.0
python-multipart==0.0.6


